Cleaned repository of duplicated files
