+++ map-viewer/viewer/index.html
@@ -29,6 +29,6 @@
     <footer>
       <p>&copy; 2024 Temperature Visualization Project CZU. All rights reserved.</p>
     </footer>
-    <script type="module" src="./map-viewer2.js"></script>
+    <script type="module" src="./map-viewer.js"></script>
   </body>
 </html>
